userspace/nat: split static-NAT compiler into nat_static.go
Continues the #4421 (codex-173 #5) pure code-motion split. Moves the
static-NAT declarations verbatim out of nat.go into nat_static.go:
clampPort (the u16 port coercion used only by the static builder) and
buildStaticNATSnapshots.

No body edits, no renames. clampPort is type-local to the static path
(its only callers are the MatchDestinationPort/MappedPort fields of the
static snapshot), so it moves with buildStaticNATSnapshots rather than
staying in the shared residual. natCounterID stays shared in nat.go.

Validation: go build ./... green.
